\name{smplab-package}
\alias{smplab-package}
\alias{smplab}
\docType{package}
\title{
\packageTitle{smplab}
}
\description{
\packageDescription{smplab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{smplab}
\packageIndices{smplab}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{smplab}

Maintainer: \packageMaintainer{smplab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
